Renamed course folder
