+++ src/components/course_content/courseBanner.jsx
