+++ src/components/course_content/courseDetails.jsx
