+++ src/components/course_content/css/courseBanner.css
