+++ src/components/course_content/css/courseDetails.css
