+++ src/components/course_content/css/relatedCourses.css
