+++ src/components/course_content/css/relatedCoursesCard.css
