+++ src/components/course_content/footer.jsx
