+++ src/components/course_content/relatedCourseCard.jsx
